app/requirements.txt
Django==4.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
django-taggit==2.1.0
django-markdownx>=4.0.0b1
django-extensions==3.1.5
Pygments==2.11.2